DRAFT Abandoning cached DOM: see below for details
Running into an unexpected problem that is making this feature
seem more complicated than it's worth (for now). The idea
was as follows:
- Highlight the "real" DOM,
- Save DOM's html into react state,
- Interweave component updates based on new react state.

The problem is that React is for some reason double rendering
some of the highlights. I suspect it is because React is comparing
new vs old state, seeing there is a new highlight tag, and then
rendering the diff. But because I've already added a highlight
tag, it double renders.