Add tiptap editor in react-hook-form
Had to use a few sources to get it working.

First was <Controller> documentation by RHF.
Source: https://react-hook-form.com/api/usecontroller/controller/

Second was answer here on StackOverflow.
Source: https://stackoverflow.com/a/71795447/

Confusing because RHF docs are not very helpful and neither was video.
The StackOverflow answer highlighted the need to set Editor's
onUpdate method to use RHF's onChange function. Neither of which
was clear in RHF or the TipTap docs.