WIP refactor to combine highlightSelection calls
We noticed there were two ways of highlightSelection is called.

The first is when annotations are fetched from the backend. The second
is when a user selects text.

Instead of having these two use cases, we can make the latter behave
like the former. When a user selects text, it sends a POST request with
a new annotation. We then invalidate the query, causing a refetch which
then highlights the text.

In essense, we've combined the two ways of calling highlightSelection.