EOD incomplete fix to annotations
Depending on whether I pass virtualDom or props.html to Interweave, I
either get React rendering bugs when I go from highlighted section to
another section, or from another section to a highlighter section,
respectively.

Possible interpretation of the first is that we apply annotations on the
old document before it has time to swap.

Possible interpretation of the second is that we swap the document
before clearHighlight() cleanup function has had time to call.