WIP fix bug where bookmark renders too early
This solves the problem of rangy functions not calling into Article is
ready, but there's now a separate issue of the bookmark not rendering
even after Article is ready (this happens randomly).

i.e. this solution stops Rangy from calling before Article is ready, but
there is now a separate issue where once Article is ready, Rangy isn't
fired since useEffect dependency on useRef is unsafe [1].

[1]  https://stackoverflow.com/q/60476155